Move root screen to app module
