+++ navigation/build.gradle.kts
@@ -1,28 +1,10 @@
 plugins {
-  id("plugin.tvmaniac.kotlin.android")
   id("plugin.tvmaniac.multiplatform")
   alias(libs.plugins.serialization)
 }
 
 kotlin {
   sourceSets {
-    androidMain.dependencies {
-      implementation(projects.android.designsystem)
-      implementation(projects.android.resources)
-
-      implementation(projects.android.feature.discover)
-      implementation(projects.android.feature.library)
-      implementation(projects.android.feature.moreShows)
-      implementation(projects.android.feature.search)
-      implementation(projects.android.feature.seasonDetails)
-      implementation(projects.android.feature.settings)
-      implementation(projects.android.feature.showDetails)
-      implementation(projects.android.feature.trailers)
-
-      implementation(libs.androidx.compose.material3)
-      implementation(libs.decompose.extensions.compose)
-    }
-
     commonMain.dependencies {
       implementation(projects.core.base)
       implementation(projects.core.traktAuth.api)
@@ -64,17 +46,4 @@ kotlin {
       implementation(libs.bundles.unittest)
     }
   }
-}
-
-android {
-  namespace = "com.thomaskioko.tvmaniac.navigation"
-
-  buildFeatures { compose = true }
-
-  compileOptions {
-    sourceCompatibility = JavaVersion.VERSION_17
-    targetCompatibility = JavaVersion.VERSION_17
-  }
-
-  composeOptions { kotlinCompilerExtensionVersion = libs.versions.composecompiler.get() }
 }
